+++ crates/assembly-syntax/src/ast/instruction/mod.rs
@@ -253,6 +253,7 @@ pub enum Instruction {
     MTreeMerge,
     MTreeVerify,
     MTreeVerifyWithError(ErrorMsg),
+    CryptoStream,
 
     // ----- STARK proof verification ------------------------------------------------------------
     FriExt2Fold4,