+++ libcore/tests/crypto/mod.rs
@@ -1,6 +1,7 @@
 #[cfg(feature = "std")]
 mod falcon;
 
+mod aead;
 mod blake3;
 mod circuit_evaluation;
 mod ecdsa_k256_keccak;